Various bugfixes for node failures/quitting.
When the NodeServer fails, the node is now correctly killed as well.  I've also added the notion of a stopped, but not failed node.  This basically represents a node that called its own stop() method to terminate execution.  The main difference is that we will not try to restart a stopped node automatically, although the user can do so manually.  The Emulator handles this by calling the Emulator's stop() method and terminating the process.