Bugfixes to 2 phase commit.  Removal of stopping without failing.
A few minor bugfixes, the most important being that 2PC now works with a non 0 coordinator, which is necessary for emulator mode, where a coordinator can come up with any address.

I've also removed the notion of stopping without a failure, because it's not necessary and greatly complicates the failure/restart model.